requirements.txt
sparse
tensorly
annoy
pandas
tqdm
spacy
hottbox
numpy
tabulate
pytorch
